Fix public crop library keyboard nav, split its page, and address code review findings
- Fix arrow-key navigation breaking after one move on the Public Crop
  Library page: navigateToLibraryState depended on location.search, so
  every selection recreated the callback chain down to loadCultures,
  re-triggering a full list reload (and losing focus) on each keypress
- Deduplicate the crop hierarchy row's expand/collapse chevron into a
  shared CropHierarchyExpandToggle component + cropHierarchyRowSx,
  used by both the desktop lists and both mobile selector dialogs
- Split PublicCropLibraryPage.tsx (2936 -> 1977 lines) into
  frontend/src/crops/components/publicCropLibrary/: formatters,
  DetailPrimitives, VersionCard, CommentForm, DiscussionComment,
  PublicCultureMobileSelectorDialog

Code review fixes (private/public culture separation branch):
- Backend: republishing an already-linked public culture now forwards
  public_variety on the update path (was silently dropped, undoing
  publish_as_general); find_owned_public_culture_for_update no longer
  skips its fallback lookup when source_public_culture is foreign-owned;
  linking to a general (variety-less) public entry no longer flags a
  false "diverged from source"; publish_as_general is resolved once via
  _resolve_public_variety instead of re-derived at each call site
- Frontend perf: CultureForm no longer rescans the full culture list on
  every keystroke in unrelated fields; PublicCropLibraryPage's seed-rate
  derivation is memoized (was recomputing every render, unlike its
  CultureDetail counterpart); the publish wizard's culture comparison is
  memoized
- Dedupe the public/private value-equality check into one
  arePublicValuesEqual (now array-order-insensitive) used by both the
  publish wizard and the culture detail view
- Fix a dead fallback in the public mobile selector's row label and a
  naive/duplicated plural-stripping search helper
- hasUnsavedChanges guard restored (was silently accepting any isDirty
  flip, not just real edits); converted its ref to state since reading
  a ref during render trips the newer react-hooks/refs rule
- Minor: type hints on new view functions, drop restated-the-obvious
  comments, drop a stray blank line

Add moderator restore for removed public cultures:
- Moderator-removed entries (status=removed) were a dead end: no
  self-service republish path by design, but also no way back at all
  since nothing exposed them past the published-only list filter
- Add PublicCultureViewSet.restore (moderator-only, no time limit) and
  reinstate_removed_public_culture; expose removed entries to
  moderators via ?status=removed instead of the default published-only
  queryset
- Add a "Removed crops" section with a restore button to
  PublicLibraryModerationPage
