+++ backend/accounts/admin.py
@@ -9,8 +9,6 @@ from .models import UserProjectSettings
 
 User = get_user_model()
 
-# Remove the default Django user admin registration so the project can provide
-# its own customized admin configuration for the user model.
 try:
     admin.site.unregister(User)
 except NotRegistered:
@@ -20,13 +18,10 @@ except NotRegistered:
 class OpenFarmPlannerUserAdmin(DjangoUserAdmin):
     """Custom admin interface configuration for the built-in Django user model."""
 
-    # Extend the default user admin columns with registration and last-login dates.
     list_display = [*DjangoUserAdmin.list_display, 'date_joined', 'last_login']
-    # Make the registration and login timestamps read-only in the admin form.
     readonly_fields = [*DjangoUserAdmin.readonly_fields, 'date_joined', 'last_login']
 
 
-# Register the customized user admin with Django's admin site.
 admin.site.register(User, OpenFarmPlannerUserAdmin)
 
 
@@ -34,7 +29,5 @@ admin.site.register(User, OpenFarmPlannerUserAdmin)
 class UserProjectSettingsAdmin(admin.ModelAdmin):
     """Admin interface configuration for the UserProjectSettings model."""
 
-    # Show the most relevant project settings fields in the admin list view.
     list_display = ['user', 'default_project', 'last_project', 'updated_at']
-    # Allow searching by the related user and project names.
     search_fields = ['user__email', 'user__username', 'default_project__name', 'last_project__name']
